Refactor new tab and fix some issues
- Move list of fortune files to file
- Use defer attribute instead of DOMContentLoaded event
- Use flexbox for centering
- Replace loops with array methods
- Replace innerHTML and HTML escaping with textContent
- Replace $.ajax with fetch and async/await
- Remove jQM
- Fix tabs being counted as one character
- Fix extra space on cow
- Fix inconsistent spacing

Libraries that are used for only a couple of things on
a page that is going to be loaded many, many times
is unneccessary.

Because we are in a Chrome extension environment we
can use the latest and greatest features without
transpiling, let's use that opportunity! (flexbox,
fetch, async/await, array methods)