Refactor and improve options
- Move option metadata to own file
- Move cow modifiers to own file
- Rename options for clarity
- Generate options from metadata
- Use flexbox for styling
- Use i18n to display human-friendly option values
- Remove normalize.css
- Replace incorrect markup

Makes options easier to maintain because all information
about them are stored in a single place.

The i18n API will make it easier to add translations and
also separates technical option values from human-friendly
values.

Normalize.css is not needed because we only have to worry
about one browser, the size of it can't be justified when
we're using this few elements.